refactor(Rank): 데이터 타입 변경
- int -> Long 으로 변경

+++ src/main/kotlin/lotto/utils/Formatter.kt
@@ -3,7 +3,7 @@ package lotto.utils
 import java.text.DecimalFormat
 
 object Formatter {
-    fun addCommas(number: Int): String {
+    fun addCommas(number: Long): String {
         val decimalFormat = DecimalFormat("#,###")
         return decimalFormat.format(number)
     }
